Reduce code duplication due to reified generics
